+++ src/ApplicationInsights/ApplicationInsights/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added breaking change warnings for upcoming Az.ApplicationInsights 2.0.0
 
 ## Version 1.3.0
 * Added WebTest function. Below is the new cmdlet